Force the name of the otp_src_.* directory to match expectations
The name of the otp_src_.* directory (e.g. otp_src_R15B03) may not
always match with the name (e.g. otp_src_R15B03-1.tar.gz) of it's
tarball file.

The tar --strip-components option was purposely not used since this
option might not be available all of the time.