Update kerl to grok KERL_INSTALL_MANPAGES env
The KERL_INSTALL_MANPAGES is ignored if given in the environment. Fixing that by saving and restoring the KERL_INSTALL_MANPAGES along the rest of the options.

By the way, KERL_INSTALL_HTMLDOCS is not reset in the beginning of the file, so no save/restore is necessary for it.

Though its an inconsistency due to be fixed... as a separate ticket.