Clean and re-run configure when --enable-native-libs is used
The OTP repository contains pre-built BEAM files that will not get
recompiled natively unless "make clean" is called first. This
commit detects when the --enable-native-libs option is used
and runs the appropriate commands.

Depending on the source of the code, this will result in the
configure step being run twice, as a Makefile is needed before
we can do "make clean". Configure will be run once otherwise.