Assert perl exists before builds (#170)
Perl 5 is required to compile Erlang from source. It's used (among other
things) to assemble BEAM opcodes, NIF jump tables and more. We now
assert perl exists on the PATH before we proceed to build Erlang or test
for Perl's version string.

We also now test for Java before we try to find its version. Java is not
required to compile Erlang though, so we don't output any warnings or
errors if its not found.