Store build configuration in a file
This change implements storing the build configuration in a file which
can later be compared to a potentially new configuration which might
trigger the need to re-run the OTP configuration steps.

We also copy the build configuration into the installation directory so
that at some future time we can write or enhance a command which shows
the build configuration.

Addresses #162 #158