Maybe munge LD and DED flags when macOS >= 10.13
As explained in #230, DED_LDFLAGS must be set correctly
for clang 9 to compile various OTP applications correctly
including crypto.

This patch reworks #230 a little bit so that it's a tiny
bit more generalized, in case we run across another
situation where this needs to be done for a particular
operating system.