Deploy kerl to pypip and install kerl from pypi
Allow user to install kerl from python pypi warehouse.

These changes introduce a python packaging system for kerl.
They package kerl into a python package, deploy it to pypi and allow
users to install kerl from pypi by using:

pip install pykerl

The kerl project name is already in use on pypi so I decided to name
it pykerl.

The package version is based on the git tag version, so when you built a
a new version, the version retrieved correspond to your tags that
already existes.

The python package and deployed is built by using travis-ci. Everythings
is automatic, you just need to deploy a new git tag to your github repo
to deploy a new version on pypi.

By example if you create a tag `1.8.6` and after you push it on github
then a new package in version `1.8.6` will be available on pypi.

You don't need to bump the version manually for the python installer part.