Pin Homebrew OpenSSL version to @1.1
Since the release of OpenSSL 3.0 Homebrew
defaults to that version, but Erlang/OTP currenly
doesn't support OpenSSL 3.0.
This either results in a build failure (when @3 is installed)
or in Erlang/OTP to be build without ssl/crypto support
(when @3 is not installed). Pinning the version to @1.1
fixes the problem, until Erlang/OTP support OpenSSL 3.0