fix: kerl upgrade always upgrading
I've noticed that calling `kerl upgrade` was always triggering an
upgrade because the local version wasn't being properly set. This was
introduced in f8c3c7a31490678b88810d75368c7ad7782c604f when the output
of `kerl version` (and the other commands) was being sent to stderr,
making the assignment `version=$(kerl version)` to be empty.

I'm not sure if changing the `tip` call to an `echo` is the proper
solution, but this would also help with external integrations that might
also want to parse the kerl version number, and that's what the included
test is testing.